move table tests to mocha and fix them
